Refactor: Moved GoalForm into goal
